alex/requirements.txt
numpy
matplotlib
imageio-ffmpeg
tqdm
